pls work

+++ scripts/setupPython.py
@@ -5,12 +5,17 @@ import importlib.util as importlib_util
 
 class PythonConfiguration:
     @classmethod
-    def Validate(cls):
+    def Validate(cls,force_install):
+
+        print("\n/*-----------------------------------------*/")
+        print("/*                   python                */")
+        print("/*-----------------------------------------*/\n")
+
         if not cls.__ValidatePython():
             return  # cannot validate further
-
+        
         for packageName in ["requests"]:
-            if not cls.__ValidatePackage(packageName):
+            if not cls.__ValidatePackage(packageName, force_install):
                 return  # cannot validate further
 
     @classmethod
@@ -36,34 +41,36 @@ class PythonConfiguration:
             return True
 
     @classmethod
-    def __ValidatePackage(cls, packageName):
+    def __ValidatePackage(cls, packageName, force_install):
         if importlib_util.find_spec(packageName) is None:
-            return cls.__InstallPackage(packageName)
+            return cls.__InstallPackage(packageName, force_install)
         return True
 
     @classmethod
-    def __InstallPackage(cls, packageName):
-        permissionGranted = False
-        while not permissionGranted:
-            reply = (
-                str(
-                    input(
-                        "Would you like to install Python package '{0:s}'? [Y/N]: ".format(
-                            packageName
+    def __InstallPackage(cls, packageName, force_install):
+        if not force_install:
+            permissionGranted = False
+            while not permissionGranted:
+                reply = (
+                    str(
+                        input(
+                            "Would you like to install Python package '{0:s}'? [Y/N]: ".format(
+                                packageName
+                            )
                         )
                     )
+                    .lower()
+                    .strip()[:1]
                 )
-                .lower()
-                .strip()[:1]
-            )
-            if reply == "n":
-                return False
-            permissionGranted = reply == "y"
-
+                if reply == "n":
+                    return False
+                permissionGranted = reply == "y"
+        else:
+            print("force installing package")
         print(f"Installing {packageName} module...")
         subprocess.check_call(["python", "-m", "pip", "install", packageName])
 
-        return cls.__ValidatePackage(packageName)
+        return cls.__ValidatePackage(packageName, force_install)
 
 
 if __name__ == "__main__":
